+++ FraudProtection/src/main/java/com/paypal/android/fraudprotection/PayPalDataCollector.kt
@@ -35,12 +35,33 @@ class PayPalDataCollector internal constructor(
      *
      * @return clientMetadataId Your server will send this to PayPal
      */
+    @Deprecated("This method is no longer supported.")
     @JvmOverloads
     fun collectDeviceData(
         context: Context,
         clientMetadataId: String? = null,
         additionalData: HashMap<String, String>? = null
     ): String {
+        val request = PayPalDataCollectorRequest(
+            hasUserLocationConsent = false,
+            clientMetadataId = clientMetadataId,
+            additionalData = additionalData
+        )
+        return collectDeviceData(context, request)
+    }
+
+    /**
+     * Use to collects device data at the time of payment. Once a user initiates a payment
+     * from their device, PayPal uses the Client Metadata ID to verify that the payment is
+     * originating from a valid, user-consented device and application. This helps reduce fraud and
+     * decrease declines. This method MUST be called prior to initiating a pre-consented payment (a
+     * "future payment") from a mobile device. Pass the result to your server, to include in the
+     * payment request sent to PayPal. Do not otherwise cache or store this value.
+     *
+     * @param context Android Context
+     * @param request Request object containing parameters to configure data collection
+     */
+    fun collectDeviceData(context: Context, request: PayPalDataCollectorRequest): String {
         val appContext = context.applicationContext
         return try {
             val magnesSettingsBuilder = MagnesSettings.Builder(appContext)
@@ -48,11 +69,12 @@ class PayPalDataCollector internal constructor(
                 .disableBeacon(false)
                 .setMagnesEnvironment(environment)
                 .setAppGuid(uuidHelper.getInstallationGUID(context))
+                .setHasUserLocationConsent(request.hasUserLocationConsent)
             magnesSDK.setUp(magnesSettingsBuilder.build())
             val result = magnesSDK.collectAndSubmit(
                 appContext,
-                clientMetadataId,
-                additionalData
+                request.clientMetadataId,
+                HashMap(request.additionalData ?: emptyMap())
             )
             result.paypalClientMetaDataId
         } catch (e: InvalidInputException) {
@@ -66,6 +88,7 @@ class PayPalDataCollector internal constructor(
         }
     }
 
+    // NEXT MAJOR VERSION: consider removing this method; it has no merchant facing purpose
     fun setLogging(shouldLog: Boolean) {
         System.setProperty("magnes.debug.mode", shouldLog.toString())
     }