+++ CHANGELOG.md
@@ -15,7 +15,15 @@
   * Add `CardAuthChallenge` type
   * Add `CardClient.removeObservers()` method
 * FraudDetection
-  * Bump Magnes dependency to version 5.4.0
+  * Fixes Google Play Store Rejection
+    * Bump Magnes version to 5.5.1
+    * Create `PayPalDataCollectorRequest`
+    * Add `PayPalDataCollector#collectDeviceData(context, request)`
+    * Deprecate `PayPalDataCollector#collectDeviceData(context, clientMetadataId, additionalData)`
+* PayPalNativePayments
+  * Fixes Google Play Store Rejection
+    * Bump Native Checkout version to 1.3.2
+    * Add `hasUserLocationConsent` to `PayPalNativeCheckoutRequest`
 * PaymentButtons
   * Update font typeface to "PayPalOpen" to meet brand guidelines
 * PayPalWebPayments